requirements.txt
alabaster==0.7.8
apturl==0.5.2
asn1crypto==0.24.0
astroid==1.6.0
Babel==2.4.0
beautifulsoup4==4.6.0
bleach==2.1.2
blinker==1.4
Brlapi==0.6.6
certifi==2018.4.16
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
decorator==4.1.2
defer==1.0.6
distro-info==0.18
dlib==19.13.1
docutils==0.14
dropbox==8.9.0
entrypoints==0.2.3.post1
face-recognition==1.2.2
face-recognition-models==0.3.0
feedparser==5.2.1
Flask==1.0.2
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
gunicorn==19.7.1
html5lib==0.999999999
httplib2==0.9.2
idna==2.7
imagesize==0.7.1
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
keyboard==0.13.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
logilab-common==1.4.1
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
numpy==1.14.5
numpydoc==0.7.0
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
opencv-contrib-python==3.4.1.15
opencv-python==3.4.1.15
pandas==0.23.1
pandocfilters==1.4.2
parso==0.1.1
passlib==1.7.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
protobuf==3.0.0
psutil==5.4.2
psycopg2==2.7.4
pycairo==1.16.2
pycodestyle==2.3.1
pycrypto==2.6.1
pycups==1.9.73
pyflakes==1.6.0
Pygments==2.2.0
pygobject==3.26.1
PyJWT==1.6.4
pylint==1.8.3
pymacaroons==0.13.0
pymongo==3.7.0
PyNaCl==1.1.2
pyRFC3339==1.0
PySocks==1.6.8
python-apt==1.6.1
python-dateutil==2.7.3
python-debian==0.1.32
pytz==2018.5
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
reportlab==3.4.0
requests==2.19.1
requests-unixsocket==0.1.5
roman==2.0.0
rope==0.10.5
RPi.GPIO==0.6.3
scikit-learn==0.19.1
scipy==1.1.0
SecretStorage==2.3.1
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
sklearn==0.0
Sphinx==1.6.7
spyder==3.2.6
SQLAlchemy==1.2.9
system-service==0.3
systemd-python==234
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
twilio==6.14.9
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.23
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.14.1
Whoosh==2.7.4
wrapt==1.9.0
WTForms==2.2.1
xkit==0.0.0
zope.interface==4.3.2
